requirements.txt
asgiref==3.4.1
beautifulsoup4==4.7.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.6
django-environ==0.9.0
django-extensions==3.1.3
django-tinymce==3.5.0
docutils==0.14
environ==1.0
gunicorn==20.1.0
idna==3.3
numpy==1.23.0
Pillow==9.2.0
psycopg2==2.9.3
Pygments==2.12.0
python-decouple==3.5
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
